requirements.txt
# SeedBox Control Panel - Backend Dependencies

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3

# Real-time WebSocket support
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0

# Docker Management
docker==7.0.0

# System Monitoring
psutil==5.9.8

# API requests for service integrations
requests==2.31.0

# WSGI Server
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.23
alembic==1.13.0

# Authentication & Security
PyJWT==2.8.1
Werkzeug==3.0.1

# Data validation
marshmallow==3.20.1
Flask-RESTful==0.3.10

# Environment
python-dotenv==1.0.0

# Monitoring & logging
structlog==24.1.0
prometheus-client==0.19.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
pylint==3.0.3
